Add simple documentation in comments, refactor.
Change all `unsigned int` and `unsigned long` types to `size_t`.
Fix names alignment in headers.

Add documentation and simple description:
Painter - painter.h
Main - tetris.c
Unicode - unicode.h
IDList - idlist.h
Engine - engine.c

Minor changes:
tetris.c - fix indentation and code blocks separation with newlines,
remove unused includes.
idlist.h - fix structure field name alignment.
field.h, engine.c - define aliases for ghost and active shapes indexes
in the field.
engine.c - rename menuTick to snake case, fix curly braces style of
functions.
Makefile - switch SFML deprecated warnings off.

+++ include/common.h
@@ -1,14 +1,14 @@
 struct game {
-    unsigned int level;
-    int          started;
-    int          paused;
-    int          scoreCurrent;
-    int          moveLatency;
-    int          lines;
-    sfClock     *gameTick;
-    sfClock     *putTick;
-    sfClock     *mTick;
-    sfClock     *repPushDown;  // repeat latency when hold Down arrow
-    sfClock     *repKeyLeft;   // repeat latency when hold Left arrow
-    sfClock     *repKeyRight;  // repeat latency when hold Right arrow
+    size_t      level;
+    int         started;
+    int         paused;
+    int         scoreCurrent;
+    int         moveLatency;
+    int         lines;
+    sfClock    *gameTick;
+    sfClock    *putTick;
+    sfClock    *mTick;
+    sfClock    *repPushDown;  // repeat latency when hold Down arrow
+    sfClock    *repKeyLeft;   // repeat latency when hold Left arrow
+    sfClock    *repKeyRight;  // repeat latency when hold Right arrow
 };
